Fix tests that fail when cache is enabled
Customers have cache enabled most of the time
Cloud customers also have them enabled

However, our tests were testing with a setup which did not include cache

This commit adds more resiliency when faced with a cache-enabled
cluster.
It validates that our changes were applied before proceeding to store
the state on our side.

This was validated by enabling cache on every test

We also had to add some validations when creating the test setup, to
make sure we wait for the cache refresh, before starting to interact
with the cluster using the provider

In the end, we removed the cache to not slow down the tests too much,
however this decision can be revisited in the future to not shoot in the
foot in future features.